Move memory file system file into src folder
